test(playwright): move role-based access tests and remove legacy KnowledgeCenter specs
Move the three article role-based access tests (ViewAll, Data Consumer,
Data Steward) from KnowledgeCenterRoleBasedAccess.spec.ts into a new
'Article Role-Based Access (Standard Roles)' describe block inside
ContextCenterPermission.spec.ts. Adapts navigation to navigateToArticles,
replaces stale testIds (add-knowledge-page-btn → create-knowledge-page-btn,
edit-owner → edit-owner-btn), reuses the existing viewOnlyPage fixture,
and replaces .ant-skeleton-active waits with waitForAllLoadersToDisappear.

Delete all four legacy KnowledgeCenter spec files now that their coverage
is fully consolidated:
- KnowledgeCenter.spec.ts
- KnowledgeCenterList.spec.ts
- KnowledgeCenterTextEditor.spec.ts
- KnowledgeCenterRoleBasedAccess.spec.ts

KnowledgeCenterTextEditor.common.ts is retained as it is still imported
by ContextCenterArticles.spec.ts.
